requirements.txt
fastapi>=0.95.0
uvicorn>=0.21.1
pydantic>=1.10.7
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
httpx>=0.24.0
neo4j>=5.8.1
redis>=4.5.5
openai>=0.27.8
python-dotenv>=1.0.0
sqlalchemy>=2.0.15
psycopg2-binary>=2.9.6 